ref: split HangTracker into three-tier architecture
Introduce a layered design for hang detection:

- RunLoopDelayTracker (renamed from DefaultHangTracker): raw
  ~25ms runloop polling via CFRunLoopObserver + semaphore
- HangTracker (new middle layer): debounces delays into hangs
  with per-observer thresholds, notifies once on threshold
  crossing and once on hang end
- WatchdogTerminationTrackingIntegration: now uses
  HangTracker.addObserver(threshold:) instead of manually
  filtering by duration
