+++ Tests/SentryTests/Integrations/ANR/HangTrackerTests.swift
@@ -2,453 +2,218 @@
 @_spi(Private) import SentryTestUtils
 import XCTest
 
-struct TestRunLoopObserver: RunLoopObserver { }
+private class MockRunLoopDelayTracker: RunLoopDelayTracker {
+    private var observers: [UUID: (TimeInterval, Bool) -> Void] = [:]
+
+    func addObserver(handler: @escaping (TimeInterval, Bool) -> Void) -> UUID {
+        let id = UUID()
+        observers[id] = handler
+        return id
+    }
+
+    func removeObserver(id: UUID) {
+        observers.removeValue(forKey: id)
+    }
+
+    func simulateDelay(duration: TimeInterval, ongoing: Bool) {
+        for observer in observers.values {
+            observer(duration, ongoing)
+        }
+    }
+}
 
 final class HangTrackerTests: XCTestCase {
-    
-    private var createdObservationBlock: ((TestRunLoopObserver?, CFRunLoopActivity) -> Void)?
-    private var observationBlock: ((TestRunLoopObserver?, CFRunLoopActivity) -> Void)?
-    private var testObserver = TestRunLoopObserver()
-    private var calledRemoveObserver = false
-    private var calledAddObserver = false
-    private let queue = DispatchQueue(label: "io.sentry.test-queue")
-    
-    override func setUp() {
-        super.setUp()
-        observationBlock = nil
-        calledRemoveObserver = false
-        calledAddObserver = false
-    }
-    
-    private func createObserver(_ allocator: CFAllocator?, _ activities: CFOptionFlags, _ repeats: Bool, _ order: CFIndex, _ block: ((TestRunLoopObserver?, CFRunLoopActivity) -> Void)?) -> TestRunLoopObserver {
-        createdObservationBlock = block
-        return testObserver
-    }
-    
-    private func addObserver(_ rl: CFRunLoop?, _ observer: TestRunLoopObserver?, _ mode: CFRunLoopMode?) {
-        observationBlock = createdObservationBlock
-        calledAddObserver = true
-    }
-    
-    private func removeObserver(_ rl: CFRunLoop?, _ observer: TestRunLoopObserver?, _ mode: CFRunLoopMode?) {
-        observationBlock = nil
-        calledRemoveObserver = true
-    }
-  
-  func testHangTrackerCallsRemoveObserverOnDealloc() {
-      var sut: DefaultHangTracker? = DefaultHangTracker(
-        dateProvider: TestCurrentDateProvider(),
-        createObserver: createObserver,
-        addObserver: addObserver,
-        removeObserver: removeObserver,
-        queue: queue)
-    _ = sut?.addOngoingHangObserver(handler: { _, _ in })
-      XCTAssertEqual(calledRemoveObserver, false)
-      sut = nil
-      XCTAssertEqual(calledRemoveObserver, true)
-  }
-    
-    func testDoesNotCaptureHangsThatAreNotOngoing() {
-        let dateProvider = TestCurrentDateProvider()
-        dateProvider.setSystemUptime(0)
-        let sut = DefaultHangTracker(
-            dateProvider: dateProvider,
-            createObserver: createObserver,
-            addObserver: addObserver,
-            removeObserver: removeObserver,
-            queue: queue)
-        
-        var observedHang = false
-        let id = sut.addOngoingHangObserver { _, _ in
-            observedHang = true
-        }
-        XCTAssertTrue(calledAddObserver, "Expected add observer to be called")
-        
-        // Ensure the queue does not run until after the full runloop
-        queue.suspend()
-        observationBlock?(testObserver, .afterWaiting)
-        // 10s passed, this is a hang
-        dateProvider.setSystemUptime(10)
-        observationBlock?(testObserver, .beforeWaiting)
-        
-        // Start the queue again
-        queue.resume()
-        
-        // This kind of hang is not caught, the hang observer is only called if
-        // a hang was caught while it is *ongoing*. Sometimes we will only know
-        // if a hang occurs after it's ended. It would be straightforward to add
-        // support for that if we ever wanted it. But for now that API isn't needed.
-        // It is best to keep that a separate API when we add it to make it clear which
-        // thread the block gets called on.
-        XCTAssertFalse(observedHang, "Should not observe hang")
-        
-        sut.removeObserver(id: id)
-        XCTAssertTrue(calledRemoveObserver, "Expected observer to be removed")
-    }
-    
-    func testHangTrackerWhenNotHanging() {
-        let dateProvider = TestCurrentDateProvider()
-        dateProvider.setSystemUptime(0)
-        let sut = DefaultHangTracker(
-            dateProvider: dateProvider,
-            createObserver: createObserver,
-            addObserver: addObserver,
-            removeObserver: removeObserver,
-            queue: queue)
-        
-        var observedHang = false
-        let id = sut.addOngoingHangObserver { _, _ in
-            observedHang = true
-        }
-        XCTAssertTrue(calledAddObserver, "Expected add observer to be called")
-        
-        // Ensure the queue does not run until after the full runloop
-        queue.suspend()
-        observationBlock?(testObserver, .afterWaiting)
-        // 10 ms passed
-        dateProvider.setSystemUptime(0.01)
-        observationBlock?(testObserver, .beforeWaiting)
-        
-        // Start the queue again
-        queue.resume()
-        
-        XCTAssertFalse(observedHang, "Should not observe hang")
-        
-        sut.removeObserver(id: id)
-        XCTAssertTrue(calledRemoveObserver, "Expected observer to be removed")
-    }
-    
-    func testHangTrackerCallsLateRunLoop() {
-        let dateProvider = TestCurrentDateProvider()
-        dateProvider.setSystemUptime(0)
-        let sut = DefaultHangTracker(
-            dateProvider: dateProvider,
-            createObserver: createObserver,
-            addObserver: addObserver,
-            removeObserver: removeObserver,
-            queue: queue)
-        
-        var observerLastInterval: TimeInterval = 0
-        var hangOngoing: Bool = false
+
+    func testObserverNotifiedWhenThresholdExceeded() {
+        let delayTracker = MockRunLoopDelayTracker()
+        let sut = DefaultHangTracker(runLoopDelayTracker: delayTracker)
+
+        var notifiedDuration: TimeInterval = 0
+        var notifiedOngoing: Bool = false
         let expectation = XCTestExpectation()
-        let id = sut.addOngoingHangObserver { interval, ongoing in
-            observerLastInterval = interval
-            hangOngoing = ongoing
+
+        let id = sut.addObserver(threshold: 0.25) { duration, ongoing in
+            notifiedDuration = duration
+            notifiedOngoing = ongoing
             expectation.fulfill()
         }
-        XCTAssertTrue(calledAddObserver, "Expected add observer to be called")
-        
-        observationBlock?(testObserver, .afterWaiting)
-        dateProvider.setSystemUptime(10)
-                
-        wait(for: [expectation])
-        
-        // Note: We are writing to these variables on a bg thread but reading them here
-        // on the main thread. This is safe without any locks because in our test
-        // environment we know that there will not be any more modifications
-        XCTAssertEqual(10, observerLastInterval, "Expected hang interval to be 10")
-        XCTAssertTrue(hangOngoing)
-        
-        observationBlock?(testObserver, .beforeWaiting)
-        
-        let expectation2 = XCTestExpectation()
-        queue.async {
-            expectation2.fulfill()
-        }
-        wait(for: [expectation2])
-        XCTAssertFalse(hangOngoing)
-        
+
+        delayTracker.simulateDelay(duration: 0.3, ongoing: true)
+
+        wait(for: [expectation], timeout: 1)
+        XCTAssertEqual(notifiedDuration, 0.3)
+        XCTAssertTrue(notifiedOngoing)
+
         sut.removeObserver(id: id)
-
-        XCTAssertTrue(calledRemoveObserver, "Expected observer to be removed")
     }
-    
-    func testRemovesObserverDuringRunloop() {
-        let dateProvider = TestCurrentDateProvider()
-        dateProvider.setSystemUptime(0)
-        let sut = DefaultHangTracker(
-            dateProvider: dateProvider,
-            createObserver: createObserver,
-            addObserver: addObserver,
-            removeObserver: removeObserver,
-            queue: queue)
-        
-        let id = sut.addOngoingHangObserver { _, _ in }
-        observationBlock?(testObserver, .afterWaiting)
+
+    func testObserverNotNotifiedWhenBelowThreshold() {
+        let delayTracker = MockRunLoopDelayTracker()
+        let sut = DefaultHangTracker(runLoopDelayTracker: delayTracker)
+
+        var notified = false
+        let id = sut.addObserver(threshold: 0.25) { _, _ in
+            notified = true
+        }
+
+        delayTracker.simulateDelay(duration: 0.2, ongoing: true)
+
+        XCTAssertFalse(notified)
+
         sut.removeObserver(id: id)
-        
-        XCTAssertTrue(calledRemoveObserver, "Expected runloop to not be observed after last observer is removed")
-        // Ensure the background queue isn't stuck waiting for another runloop event
-        let expectation = XCTestExpectation()
-        queue.async {
-            expectation.fulfill()
-        }
-        // Ensure the queue is not blocked
-        wait(for: [expectation])
     }
-    
-    func testHangTrackerDeallocates() {
-        let dateProvider = TestCurrentDateProvider()
-        dateProvider.setSystemUptime(0)
-        var sut: DefaultHangTracker? = DefaultHangTracker(
-            dateProvider: dateProvider,
-            createObserver: createObserver,
-            addObserver: addObserver,
-            removeObserver: removeObserver,
-            queue: queue)
-        #if swift(>=6.2)
-        weak let weakSut = sut
-        #else
-        weak var weakSut = sut
-        #endif
-        
-        _ = sut?.addOngoingHangObserver { _, _ in }
-        observationBlock?(testObserver, .afterWaiting)
-        observationBlock?(testObserver, .beforeWaiting)
-        
-        sut = nil
-        
-        // Allow the hang tracker's background thread to finish since it holds
-        // a strong reference while it is running
-        let expectation = XCTestExpectation()
-        queue.async {
-            expectation.fulfill()
-        }
-        wait(for: [expectation])
-        
-        XCTAssertNil(weakSut, "Expected observer to be deallocated")
-    }
-    
-    /// Verifies that after one hang completes (ongoing=true then ongoing=false),
-    /// a second hang is properly detected. This catches state-reset bugs with consecutive hangs.
-    func testConsecutiveHangsAreDetected() {
-        let dateProvider = TestCurrentDateProvider()
-        dateProvider.setSystemUptime(0)
-        let sut = DefaultHangTracker(
-            dateProvider: dateProvider,
-            createObserver: createObserver,
-            addObserver: addObserver,
-            removeObserver: removeObserver,
-            queue: queue)
-        
-        let lock = NSLock()
-        var hangCount = 0
-        var lastInterval: TimeInterval = 0
-        var lastOngoing: Bool = false
-        var hangCallback = XCTestExpectation()
-        let id = sut.addOngoingHangObserver { interval, ongoing in
-            // Only fulfill one time
-            lock.synchronized {
-                if lastInterval == 0 {
-                    hangCallback.fulfill()
-                }
-                lastInterval = interval
-                lastOngoing = ongoing
-                if !ongoing {
-                    hangCount += 1
-                }
-            }
-        }
-        
-        // First hang: start
-        observationBlock?(testObserver, .afterWaiting)
-        dateProvider.setSystemUptime(10)
-        wait(for: [hangCallback])
-        
-        lock.synchronized {
-            XCTAssertEqual(lastInterval, 10, "First hang interval should be 10")
-            XCTAssertTrue(lastOngoing, "First hang should be ongoing")
-        }
-        
-        // First hang: complete
-        observationBlock?(testObserver, .beforeWaiting)
-        
-        let firstHangEndExpectation = XCTestExpectation(description: "First hang ended")
-        queue.async {
-            firstHangEndExpectation.fulfill()
-        }
-        wait(for: [firstHangEndExpectation])
 
-        // No need for the lock here, the background thread finished
-        XCTAssertEqual(hangCount, 1, "First hang should be detected")
-        XCTAssertFalse(lastOngoing, "First hang should no longer be ongoing")
-        
-        // Second hang: start (simulating another runloop iteration that hangs)
-        dateProvider.setSystemUptime(20)
-        observationBlock?(testObserver, .afterWaiting)
-        dateProvider.setSystemUptime(35) // 15 second hang
+    func testObserverNotNotifiedWhenExactlyAtThreshold() {
+        let delayTracker = MockRunLoopDelayTracker()
+        let sut = DefaultHangTracker(runLoopDelayTracker: delayTracker)
 
-        lock.synchronized {
-            hangCallback = XCTestExpectation(description: "Second hang detected")
-            lastInterval = 0
+        var notified = false
+        let id = sut.addObserver(threshold: 0.25) { _, _ in
+            notified = true
         }
-        wait(for: [hangCallback])
 
-        lock.synchronized {
-            XCTAssertEqual(lastInterval, 15, "Second hang interval should be 15")
-            XCTAssertTrue(lastOngoing, "Second hang should be ongoing")
-        }
-        
-        // Second hang: complete
-        observationBlock?(testObserver, .beforeWaiting)
-        
-        let secondHangEndExpectation = XCTestExpectation(description: "Second hang ended")
-        queue.async {
-            secondHangEndExpectation.fulfill()
-        }
-        wait(for: [secondHangEndExpectation])
-        
-        // No need fo rthe lock here, the background thread finished
-        XCTAssertEqual(hangCount, 2, "Second hang should be detected after first hang completed")
-        XCTAssertFalse(lastOngoing, "Second hang should no longer be ongoing")
-        
+        delayTracker.simulateDelay(duration: 0.25, ongoing: true)
+
+        XCTAssertFalse(notified)
+
         sut.removeObserver(id: id)
-        XCTAssertTrue(calledRemoveObserver, "Expected observer to be removed")
     }
 
-    /// Verifies that when all references to HangTracker are nilled while the background queue
-    /// is still in the waitForHang loop, the class does not deallocate until the loop exits,
-    /// and then the dispatch queue is freed up (not blocked).
-    func testDeallocWhileInWaitForHangLoop() {
-        let dateProvider = TestCurrentDateProvider()
-        dateProvider.setSystemUptime(0)
-        var sut: DefaultHangTracker? = DefaultHangTracker(
-            dateProvider: dateProvider,
-            createObserver: createObserver,
-            addObserver: addObserver,
-            removeObserver: removeObserver,
-            queue: queue)
-        #if swift(>=6.2)
-        weak let weakSut = sut
-        #else
-        weak var weakSut = sut
-        #endif
-        
-        let expectation = XCTestExpectation()
-        var hangDetected = false
-        _ = sut?.addOngoingHangObserver { _, _ in
-            if !hangDetected {
-                expectation.fulfill()
-                hangDetected = true
-            }
+    func testObserverOnlyNotifiedOncePerHang() {
+        let delayTracker = MockRunLoopDelayTracker()
+        let sut = DefaultHangTracker(runLoopDelayTracker: delayTracker)
+
+        var callCount = 0
+        let id = sut.addObserver(threshold: 0.25) { _, ongoing in
+            if ongoing { callCount += 1 }
         }
-        
-        // Start the runloop iteration - this triggers the background queue to start waiting
-        observationBlock?(testObserver, .afterWaiting)
-        
-        // Wait until the hang is detected
-        wait(for: [expectation])
-        
-        // Now nil all references while the background queue is in the waitForHang loop
-        // The HangTracker should NOT immediately deallocate because the background queue
-        // holds a reference via the closure
-        sut = nil
-        
-        XCTAssertNotNil(weakSut)
-        
-        observationBlock?(testObserver, .beforeWaiting)
-        
-        // Verify the queue is not blocked
-        let queueFreeExpectation = XCTestExpectation(description: "Queue is free")
-        queue.async {
-            queueFreeExpectation.fulfill()
-        }
-        wait(for: [queueFreeExpectation])
-        XCTAssertNil(weakSut)
-        XCTAssertTrue(calledRemoveObserver)
+
+        delayTracker.simulateDelay(duration: 0.3, ongoing: true)
+        delayTracker.simulateDelay(duration: 0.5, ongoing: true)
+        delayTracker.simulateDelay(duration: 1.0, ongoing: true)
+
+        XCTAssertEqual(callCount, 1, "Observer should only be notified once per hang")
+
+        sut.removeObserver(id: id)
     }
 
-    func testMultipleObserversAllReceiveHangCallback() {
-        let dateProvider = TestCurrentDateProvider()
-        dateProvider.setSystemUptime(0)
-        let sut = DefaultHangTracker(
-            dateProvider: dateProvider,
-            createObserver: createObserver,
-            addObserver: addObserver,
-            removeObserver: removeObserver,
-            queue: queue)
+    func testObserverReceivesEndNotificationAfterThresholdCrossed() {
+        let delayTracker = MockRunLoopDelayTracker()
+        let sut = DefaultHangTracker(runLoopDelayTracker: delayTracker)
 
-        let lock = NSLock()
-        var observer1Interval: TimeInterval = 0
-        var observer1Ongoing: Bool = false
-        var observer2Interval: TimeInterval = 0
-        var observer2Ongoing: Bool = false
-        var observer3Interval: TimeInterval = 0
-        var observer3Ongoing: Bool = false
-
-        let expectation1 = XCTestExpectation(description: "Observer 1 called")
-        let expectation2 = XCTestExpectation(description: "Observer 2 called")
-        let expectation3 = XCTestExpectation(description: "Observer 3 called")
-
-        let id1 = sut.addOngoingHangObserver { interval, ongoing in
-            // Only fulfill one time
-            if observer1Interval == 0 {
-                expectation1.fulfill()
-            }
-            lock.synchronized {
-                observer1Interval = interval
-                observer1Ongoing = ongoing
-            }
-        }
-        let id2 = sut.addOngoingHangObserver { interval, ongoing in
-            // Only fulfill one time
-            if observer2Interval == 0 {
-                expectation2.fulfill()
-            }
-            lock.synchronized {
-                observer2Interval = interval
-                observer2Ongoing = ongoing
-            }
-        }
-        let id3 = sut.addOngoingHangObserver { interval, ongoing in
-            // Only fulfill one time
-            if observer3Interval == 0 {
-                expectation3.fulfill()
-            }
-            lock.synchronized {
-                observer3Interval = interval
-                observer3Ongoing = ongoing
+        var ongoingCalls = 0
+        var endedCalls = 0
+        var endedDuration: TimeInterval = 0
+        let id = sut.addObserver(threshold: 0.25) { duration, ongoing in
+            if ongoing {
+                ongoingCalls += 1
+            } else {
+                endedCalls += 1
+                endedDuration = duration
             }
         }
 
-        XCTAssertTrue(calledAddObserver, "Expected add observer to be called")
+        delayTracker.simulateDelay(duration: 0.5, ongoing: true)
+        delayTracker.simulateDelay(duration: 1.0, ongoing: false)
 
-        // Trigger a hang
-        observationBlock?(testObserver, .afterWaiting)
-        dateProvider.setSystemUptime(10)
+        XCTAssertEqual(ongoingCalls, 1)
+        XCTAssertEqual(endedCalls, 1)
+        XCTAssertEqual(endedDuration, 1.0)
 
-        wait(for: [expectation1, expectation2, expectation3])
+        sut.removeObserver(id: id)
+    }
 
-        lock.synchronized {
-            // All observers should have received the hang with same interval
-            XCTAssertEqual(observer1Interval, 10, "Observer 1 should receive hang interval")
-            XCTAssertEqual(observer2Interval, 10, "Observer 2 should receive hang interval")
-            XCTAssertEqual(observer3Interval, 10, "Observer 3 should receive hang interval")
-            
-            XCTAssertTrue(observer1Ongoing, "Observer 1 should report hang as ongoing")
-            XCTAssertTrue(observer2Ongoing, "Observer 2 should report hang as ongoing")
-            XCTAssertTrue(observer3Ongoing, "Observer 3 should report hang as ongoing")
+    func testObserverDoesNotReceiveEndNotificationWithoutPriorOngoing() {
+        let delayTracker = MockRunLoopDelayTracker()
+        let sut = DefaultHangTracker(runLoopDelayTracker: delayTracker)
+
+        var notified = false
+        let id = sut.addObserver(threshold: 2.0) { _, _ in
+            notified = true
         }
 
-        // End the hang
-        observationBlock?(testObserver, .beforeWaiting)
+        // Delay below threshold, then ended — observer should not get the end notification
+        delayTracker.simulateDelay(duration: 0.5, ongoing: true)
+        delayTracker.simulateDelay(duration: 1.0, ongoing: false)
 
-        let hangEndExpectation = XCTestExpectation(description: "Hang ended")
-        queue.async {
-            hangEndExpectation.fulfill()
+        XCTAssertFalse(notified)
+
+        sut.removeObserver(id: id)
+    }
+
+    func testMultipleObserversWithDifferentThresholds() {
+        let delayTracker = MockRunLoopDelayTracker()
+        let sut = DefaultHangTracker(runLoopDelayTracker: delayTracker)
+
+        var lowThresholdNotified = false
+        var highThresholdNotified = false
+
+        let id1 = sut.addObserver(threshold: 0.25) { _, ongoing in
+            if ongoing { lowThresholdNotified = true }
         }
-        wait(for: [hangEndExpectation])
+        let id2 = sut.addObserver(threshold: 2.0) { _, ongoing in
+            if ongoing { highThresholdNotified = true }
+        }
 
-        // No need for the lock here because the hang is over
-        // All observers should have been notified that the hang ended
-        XCTAssertFalse(observer1Ongoing, "Observer 1 should report hang ended")
-        XCTAssertFalse(observer2Ongoing, "Observer 2 should report hang ended")
-        XCTAssertFalse(observer3Ongoing, "Observer 3 should report hang ended")
+        // Duration exceeds only the low threshold
+        delayTracker.simulateDelay(duration: 0.5, ongoing: true)
+
+        XCTAssertTrue(lowThresholdNotified, "Low threshold observer should be notified")
+        XCTAssertFalse(highThresholdNotified, "High threshold observer should NOT be notified")
+
+        // Duration now exceeds both thresholds
+        delayTracker.simulateDelay(duration: 3.0, ongoing: true)
+
+        XCTAssertTrue(highThresholdNotified, "High threshold observer should now be notified")
 
         sut.removeObserver(id: id1)
         sut.removeObserver(id: id2)
-        sut.removeObserver(id: id3)
-        XCTAssertTrue(calledRemoveObserver, "Expected observer to be removed")
     }
 
+    func testConsecutiveHangsResetState() {
+        let delayTracker = MockRunLoopDelayTracker()
+        let sut = DefaultHangTracker(runLoopDelayTracker: delayTracker)
+
+        var ongoingCalls = 0
+        var endedCalls = 0
+        let id = sut.addObserver(threshold: 0.25) { _, ongoing in
+            if ongoing {
+                ongoingCalls += 1
+            } else {
+                endedCalls += 1
+            }
+        }
+
+        // First hang
+        delayTracker.simulateDelay(duration: 0.5, ongoing: true)
+        delayTracker.simulateDelay(duration: 0.8, ongoing: false)
+
+        XCTAssertEqual(ongoingCalls, 1)
+        XCTAssertEqual(endedCalls, 1)
+
+        // Second hang — should be detected again
+        delayTracker.simulateDelay(duration: 0.5, ongoing: true)
+        delayTracker.simulateDelay(duration: 1.0, ongoing: false)
+
+        XCTAssertEqual(ongoingCalls, 2, "Second hang should trigger a new ongoing notification")
+        XCTAssertEqual(endedCalls, 2, "Second hang should trigger a new ended notification")
+
+        sut.removeObserver(id: id)
+    }
+
+    func testRemoveObserverStopsNotifications() {
+        let delayTracker = MockRunLoopDelayTracker()
+        let sut = DefaultHangTracker(runLoopDelayTracker: delayTracker)
+
+        var notified = false
+        let id = sut.addObserver(threshold: 0.25) { _, _ in
+            notified = true
+        }
+
+        sut.removeObserver(id: id)
+
+        delayTracker.simulateDelay(duration: 1.0, ongoing: true)
+
+        XCTAssertFalse(notified)
+    }
 }
