+++ Tests/SentryTests/Integrations/ANR/RunLoopDelayTrackerTests.swift
@@ -1,0 +1,454 @@
+@_spi(Private) @testable import Sentry
+@_spi(Private) import SentryTestUtils
+import XCTest
+
+struct TestRunLoopObserver: RunLoopObserver { }
+
+final class RunLoopDelayTrackerTests: XCTestCase {
+    
+    private var createdObservationBlock: ((TestRunLoopObserver?, CFRunLoopActivity) -> Void)?
+    private var observationBlock: ((TestRunLoopObserver?, CFRunLoopActivity) -> Void)?
+    private var testObserver = TestRunLoopObserver()
+    private var calledRemoveObserver = false
+    private var calledAddObserver = false
+    private let queue = DispatchQueue(label: "io.sentry.test-queue")
+    
+    override func setUp() {
+        super.setUp()
+        observationBlock = nil
+        calledRemoveObserver = false
+        calledAddObserver = false
+    }
+    
+    private func createObserver(_ allocator: CFAllocator?, _ activities: CFOptionFlags, _ repeats: Bool, _ order: CFIndex, _ block: ((TestRunLoopObserver?, CFRunLoopActivity) -> Void)?) -> TestRunLoopObserver {
+        createdObservationBlock = block
+        return testObserver
+    }
+    
+    private func addObserver(_ rl: CFRunLoop?, _ observer: TestRunLoopObserver?, _ mode: CFRunLoopMode?) {
+        observationBlock = createdObservationBlock
+        calledAddObserver = true
+    }
+    
+    private func removeObserver(_ rl: CFRunLoop?, _ observer: TestRunLoopObserver?, _ mode: CFRunLoopMode?) {
+        observationBlock = nil
+        calledRemoveObserver = true
+    }
+  
+  func testHangTrackerCallsRemoveObserverOnDealloc() {
+      var sut: DefaultRunLoopDelayTracker? = DefaultRunLoopDelayTracker(
+        dateProvider: TestCurrentDateProvider(),
+        createObserver: createObserver,
+        addObserver: addObserver,
+        removeObserver: removeObserver,
+        queue: queue)
+    _ = sut?.addObserver { _, _ in }
+      XCTAssertEqual(calledRemoveObserver, false)
+      sut = nil
+      XCTAssertEqual(calledRemoveObserver, true)
+  }
+    
+    func testDoesNotCaptureHangsThatAreNotOngoing() {
+        let dateProvider = TestCurrentDateProvider()
+        dateProvider.setSystemUptime(0)
+        let sut = DefaultRunLoopDelayTracker(
+            dateProvider: dateProvider,
+            createObserver: createObserver,
+            addObserver: addObserver,
+            removeObserver: removeObserver,
+            queue: queue)
+        
+        var observedHang = false
+        let id = sut.addObserver { _, _ in
+            observedHang = true
+        }
+        XCTAssertTrue(calledAddObserver, "Expected add observer to be called")
+        
+        // Ensure the queue does not run until after the full runloop
+        queue.suspend()
+        observationBlock?(testObserver, .afterWaiting)
+        // 10s passed, this is a hang
+        dateProvider.setSystemUptime(10)
+        observationBlock?(testObserver, .beforeWaiting)
+        
+        // Start the queue again
+        queue.resume()
+        
+        // This kind of hang is not caught, the hang observer is only called if
+        // a hang was caught while it is *ongoing*. Sometimes we will only know
+        // if a hang occurs after it's ended. It would be straightforward to add
+        // support for that if we ever wanted it. But for now that API isn't needed.
+        // It is best to keep that a separate API when we add it to make it clear which
+        // thread the block gets called on.
+        XCTAssertFalse(observedHang, "Should not observe hang")
+        
+        sut.removeObserver(id: id)
+        XCTAssertTrue(calledRemoveObserver, "Expected observer to be removed")
+    }
+    
+    func testHangTrackerWhenNotHanging() {
+        let dateProvider = TestCurrentDateProvider()
+        dateProvider.setSystemUptime(0)
+        let sut = DefaultRunLoopDelayTracker(
+            dateProvider: dateProvider,
+            createObserver: createObserver,
+            addObserver: addObserver,
+            removeObserver: removeObserver,
+            queue: queue)
+        
+        var observedHang = false
+        let id = sut.addObserver { _, _ in
+            observedHang = true
+        }
+        XCTAssertTrue(calledAddObserver, "Expected add observer to be called")
+        
+        // Ensure the queue does not run until after the full runloop
+        queue.suspend()
+        observationBlock?(testObserver, .afterWaiting)
+        // 10 ms passed
+        dateProvider.setSystemUptime(0.01)
+        observationBlock?(testObserver, .beforeWaiting)
+        
+        // Start the queue again
+        queue.resume()
+        
+        XCTAssertFalse(observedHang, "Should not observe hang")
+        
+        sut.removeObserver(id: id)
+        XCTAssertTrue(calledRemoveObserver, "Expected observer to be removed")
+    }
+    
+    func testHangTrackerCallsLateRunLoop() {
+        let dateProvider = TestCurrentDateProvider()
+        dateProvider.setSystemUptime(0)
+        let sut = DefaultRunLoopDelayTracker(
+            dateProvider: dateProvider,
+            createObserver: createObserver,
+            addObserver: addObserver,
+            removeObserver: removeObserver,
+            queue: queue)
+        
+        var observerLastInterval: TimeInterval = 0
+        var hangOngoing: Bool = false
+        let expectation = XCTestExpectation()
+        let id = sut.addObserver { interval, ongoing in
+            observerLastInterval = interval
+            hangOngoing = ongoing
+            expectation.fulfill()
+        }
+        XCTAssertTrue(calledAddObserver, "Expected add observer to be called")
+        
+        observationBlock?(testObserver, .afterWaiting)
+        dateProvider.setSystemUptime(10)
+                
+        wait(for: [expectation])
+        
+        // Note: We are writing to these variables on a bg thread but reading them here
+        // on the main thread. This is safe without any locks because in our test
+        // environment we know that there will not be any more modifications
+        XCTAssertEqual(10, observerLastInterval, "Expected hang interval to be 10")
+        XCTAssertTrue(hangOngoing)
+        
+        observationBlock?(testObserver, .beforeWaiting)
+        
+        let expectation2 = XCTestExpectation()
+        queue.async {
+            expectation2.fulfill()
+        }
+        wait(for: [expectation2])
+        XCTAssertFalse(hangOngoing)
+        
+        sut.removeObserver(id: id)
+
+        XCTAssertTrue(calledRemoveObserver, "Expected observer to be removed")
+    }
+    
+    func testRemovesObserverDuringRunloop() {
+        let dateProvider = TestCurrentDateProvider()
+        dateProvider.setSystemUptime(0)
+        let sut = DefaultRunLoopDelayTracker(
+            dateProvider: dateProvider,
+            createObserver: createObserver,
+            addObserver: addObserver,
+            removeObserver: removeObserver,
+            queue: queue)
+        
+        let id = sut.addObserver { _, _ in }
+        observationBlock?(testObserver, .afterWaiting)
+        sut.removeObserver(id: id)
+        
+        XCTAssertTrue(calledRemoveObserver, "Expected runloop to not be observed after last observer is removed")
+        // Ensure the background queue isn't stuck waiting for another runloop event
+        let expectation = XCTestExpectation()
+        queue.async {
+            expectation.fulfill()
+        }
+        // Ensure the queue is not blocked
+        wait(for: [expectation])
+    }
+    
+    func testHangTrackerDeallocates() {
+        let dateProvider = TestCurrentDateProvider()
+        dateProvider.setSystemUptime(0)
+        var sut: DefaultRunLoopDelayTracker? = DefaultRunLoopDelayTracker(
+            dateProvider: dateProvider,
+            createObserver: createObserver,
+            addObserver: addObserver,
+            removeObserver: removeObserver,
+            queue: queue)
+        #if swift(>=6.2)
+        weak let weakSut = sut
+        #else
+        weak var weakSut = sut
+        #endif
+        
+        _ = sut?.addObserver { _, _ in }
+        observationBlock?(testObserver, .afterWaiting)
+        observationBlock?(testObserver, .beforeWaiting)
+        
+        sut = nil
+        
+        // Allow the hang tracker's background thread to finish since it holds
+        // a strong reference while it is running
+        let expectation = XCTestExpectation()
+        queue.async {
+            expectation.fulfill()
+        }
+        wait(for: [expectation])
+        
+        XCTAssertNil(weakSut, "Expected observer to be deallocated")
+    }
+    
+    /// Verifies that after one hang completes (ongoing=true then ongoing=false),
+    /// a second hang is properly detected. This catches state-reset bugs with consecutive hangs.
+    func testConsecutiveHangsAreDetected() {
+        let dateProvider = TestCurrentDateProvider()
+        dateProvider.setSystemUptime(0)
+        let sut = DefaultRunLoopDelayTracker(
+            dateProvider: dateProvider,
+            createObserver: createObserver,
+            addObserver: addObserver,
+            removeObserver: removeObserver,
+            queue: queue)
+        
+        let lock = NSLock()
+        var hangCount = 0
+        var lastInterval: TimeInterval = 0
+        var lastOngoing: Bool = false
+        var hangCallback = XCTestExpectation()
+        let id = sut.addObserver { interval, ongoing in
+            // Only fulfill one time
+            lock.synchronized {
+                if lastInterval == 0 {
+                    hangCallback.fulfill()
+                }
+                lastInterval = interval
+                lastOngoing = ongoing
+                if !ongoing {
+                    hangCount += 1
+                }
+            }
+        }
+        
+        // First hang: start
+        observationBlock?(testObserver, .afterWaiting)
+        dateProvider.setSystemUptime(10)
+        wait(for: [hangCallback])
+        
+        lock.synchronized {
+            XCTAssertEqual(lastInterval, 10, "First hang interval should be 10")
+            XCTAssertTrue(lastOngoing, "First hang should be ongoing")
+        }
+        
+        // First hang: complete
+        observationBlock?(testObserver, .beforeWaiting)
+        
+        let firstHangEndExpectation = XCTestExpectation(description: "First hang ended")
+        queue.async {
+            firstHangEndExpectation.fulfill()
+        }
+        wait(for: [firstHangEndExpectation])
+
+        // No need for the lock here, the background thread finished
+        XCTAssertEqual(hangCount, 1, "First hang should be detected")
+        XCTAssertFalse(lastOngoing, "First hang should no longer be ongoing")
+        
+        // Second hang: start (simulating another runloop iteration that hangs)
+        dateProvider.setSystemUptime(20)
+        observationBlock?(testObserver, .afterWaiting)
+        dateProvider.setSystemUptime(35) // 15 second hang
+
+        lock.synchronized {
+            hangCallback = XCTestExpectation(description: "Second hang detected")
+            lastInterval = 0
+        }
+        wait(for: [hangCallback])
+
+        lock.synchronized {
+            XCTAssertEqual(lastInterval, 15, "Second hang interval should be 15")
+            XCTAssertTrue(lastOngoing, "Second hang should be ongoing")
+        }
+        
+        // Second hang: complete
+        observationBlock?(testObserver, .beforeWaiting)
+        
+        let secondHangEndExpectation = XCTestExpectation(description: "Second hang ended")
+        queue.async {
+            secondHangEndExpectation.fulfill()
+        }
+        wait(for: [secondHangEndExpectation])
+        
+        // No need fo rthe lock here, the background thread finished
+        XCTAssertEqual(hangCount, 2, "Second hang should be detected after first hang completed")
+        XCTAssertFalse(lastOngoing, "Second hang should no longer be ongoing")
+        
+        sut.removeObserver(id: id)
+        XCTAssertTrue(calledRemoveObserver, "Expected observer to be removed")
+    }
+
+    /// Verifies that when all references to HangTracker are nilled while the background queue
+    /// is still in the waitForHang loop, the class does not deallocate until the loop exits,
+    /// and then the dispatch queue is freed up (not blocked).
+    func testDeallocWhileInWaitForHangLoop() {
+        let dateProvider = TestCurrentDateProvider()
+        dateProvider.setSystemUptime(0)
+        var sut: DefaultRunLoopDelayTracker? = DefaultRunLoopDelayTracker(
+            dateProvider: dateProvider,
+            createObserver: createObserver,
+            addObserver: addObserver,
+            removeObserver: removeObserver,
+            queue: queue)
+        #if swift(>=6.2)
+        weak let weakSut = sut
+        #else
+        weak var weakSut = sut
+        #endif
+        
+        let expectation = XCTestExpectation()
+        var hangDetected = false
+        _ = sut?.addObserver { _, _ in
+            if !hangDetected {
+                expectation.fulfill()
+                hangDetected = true
+            }
+        }
+        
+        // Start the runloop iteration - this triggers the background queue to start waiting
+        observationBlock?(testObserver, .afterWaiting)
+        
+        // Wait until the hang is detected
+        wait(for: [expectation])
+        
+        // Now nil all references while the background queue is in the waitForHang loop
+        // The HangTracker should NOT immediately deallocate because the background queue
+        // holds a reference via the closure
+        sut = nil
+        
+        XCTAssertNotNil(weakSut)
+        
+        observationBlock?(testObserver, .beforeWaiting)
+        
+        // Verify the queue is not blocked
+        let queueFreeExpectation = XCTestExpectation(description: "Queue is free")
+        queue.async {
+            queueFreeExpectation.fulfill()
+        }
+        wait(for: [queueFreeExpectation])
+        XCTAssertNil(weakSut)
+        XCTAssertTrue(calledRemoveObserver)
+    }
+
+    func testMultipleObserversAllReceiveHangCallback() {
+        let dateProvider = TestCurrentDateProvider()
+        dateProvider.setSystemUptime(0)
+        let sut = DefaultRunLoopDelayTracker(
+            dateProvider: dateProvider,
+            createObserver: createObserver,
+            addObserver: addObserver,
+            removeObserver: removeObserver,
+            queue: queue)
+
+        let lock = NSLock()
+        var observer1Interval: TimeInterval = 0
+        var observer1Ongoing: Bool = false
+        var observer2Interval: TimeInterval = 0
+        var observer2Ongoing: Bool = false
+        var observer3Interval: TimeInterval = 0
+        var observer3Ongoing: Bool = false
+
+        let expectation1 = XCTestExpectation(description: "Observer 1 called")
+        let expectation2 = XCTestExpectation(description: "Observer 2 called")
+        let expectation3 = XCTestExpectation(description: "Observer 3 called")
+
+        let id1 = sut.addObserver { interval, ongoing in
+            // Only fulfill one time
+            if observer1Interval == 0 {
+                expectation1.fulfill()
+            }
+            lock.synchronized {
+                observer1Interval = interval
+                observer1Ongoing = ongoing
+            }
+        }
+        let id2 = sut.addObserver { interval, ongoing in
+            // Only fulfill one time
+            if observer2Interval == 0 {
+                expectation2.fulfill()
+            }
+            lock.synchronized {
+                observer2Interval = interval
+                observer2Ongoing = ongoing
+            }
+        }
+        let id3 = sut.addObserver { interval, ongoing in
+            // Only fulfill one time
+            if observer3Interval == 0 {
+                expectation3.fulfill()
+            }
+            lock.synchronized {
+                observer3Interval = interval
+                observer3Ongoing = ongoing
+            }
+        }
+
+        XCTAssertTrue(calledAddObserver, "Expected add observer to be called")
+
+        // Trigger a hang
+        observationBlock?(testObserver, .afterWaiting)
+        dateProvider.setSystemUptime(10)
+
+        wait(for: [expectation1, expectation2, expectation3])
+
+        lock.synchronized {
+            // All observers should have received the hang with same interval
+            XCTAssertEqual(observer1Interval, 10, "Observer 1 should receive hang interval")
+            XCTAssertEqual(observer2Interval, 10, "Observer 2 should receive hang interval")
+            XCTAssertEqual(observer3Interval, 10, "Observer 3 should receive hang interval")
+            
+            XCTAssertTrue(observer1Ongoing, "Observer 1 should report hang as ongoing")
+            XCTAssertTrue(observer2Ongoing, "Observer 2 should report hang as ongoing")
+            XCTAssertTrue(observer3Ongoing, "Observer 3 should report hang as ongoing")
+        }
+
+        // End the hang
+        observationBlock?(testObserver, .beforeWaiting)
+
+        let hangEndExpectation = XCTestExpectation(description: "Hang ended")
+        queue.async {
+            hangEndExpectation.fulfill()
+        }
+        wait(for: [hangEndExpectation])
+
+        // No need for the lock here because the hang is over
+        // All observers should have been notified that the hang ended
+        XCTAssertFalse(observer1Ongoing, "Observer 1 should report hang ended")
+        XCTAssertFalse(observer2Ongoing, "Observer 2 should report hang ended")
+        XCTAssertFalse(observer3Ongoing, "Observer 3 should report hang ended")
+
+        sut.removeObserver(id: id1)
+        sut.removeObserver(id: id2)
+        sut.removeObserver(id: id3)
+        XCTAssertTrue(calledRemoveObserver, "Expected observer to be removed")
+    }
+
+}
